pyproject.toml
[project]
name = "module-2"
version = "0.1.0"
description = "A command line chatbot capable of streaming text"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
